+++ src/edu/tongji/sse/model/Line.java
@@ -4,10 +4,10 @@ package edu.tongji.sse.model;
  * Created by huage on 2017/4/16.
  */
 public class Line {
-    public Integer lineHash;
+    public Long lineHash;
     public int lineNum;
 
-    public Line(Integer lineHash, int lineNum) {
+    public Line(Long lineHash, int lineNum) {
         this.lineHash = lineHash;
         this.lineNum = lineNum;
     }
